server/requirements.txt
flask
langchain
llama-index==0.6.9
weaviate-client
pymongo
flask_cors
flask_socketio
eventlet